+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Bug fixes:
 
+- Fix unspecified layer warning being emitted reading multilayer datasets, even
+  when layer is specified when using the mask or bbox keywords (#3378).
 - Properly support named aggregations over a geometry column in `GroupBy.agg` (#3368).
 - Fix regression where constructing a GeoSeries from a pd.Series with GeometryDtype values 
   failed when `crs` was provided(#3383).